With Differentiable Argmax + FH tensorflow version
