+++ PretrainedModels/patternModel.json
@@ -1,0 +1,1 @@
+{"class_name": "Model", "config": {"name": "model_2", "layers": [{"class_name": "InputLayer", "config": {"batch_input_shape": [null, 256, 1], "dtype": "float32", "sparse": false, "ragged": false, "name": "hist"}, "name": "hist", "inbound_nodes": []}, {"class_name": "Conv1D", "config": {"name": "pattern_1", "trainable": true, "dtype": "float32", "filters": 16, "kernel_size": [4], "strides": [1], "padding": "same", "data_format": "channels_last", "dilation_rate": [1], "activation": "relu", "use_bias": true, "kernel_initializer": {"class_name": "VarianceScaling", "config": {"scale": 1.0, "mode": "fan_in", "distribution": "truncated_normal", "seed": null}}, "bias_initializer": {"class_name": "Zeros", "config": {}}, "kernel_regularizer": {"class_name": "L1L2", "config": {"l1": 0.0, "l2": 1.000000013351432e-10}}, "bias_regularizer": null, "activity_regularizer": null, "kernel_constraint": null, "bias_constraint": null}, "name": "pattern_1", "inbound_nodes": [[["hist", 0, 0, {}]]]}, {"class_name": "Conv1D", "config": {"name": "pattern_2", "trainable": true, "dtype": "float32", "filters": 16, "kernel_size": [4], "strides": [1], "padding": "same", "data_format": "channels_last", "dilation_rate": [1], "activation": "relu", "use_bias": true, "kernel_initializer": {"class_name": "VarianceScaling", "config": {"scale": 1.0, "mode": "fan_in", "distribution": "truncated_normal", "seed": null}}, "bias_initializer": {"class_name": "Zeros", "config": {}}, "kernel_regularizer": {"class_name": "L1L2", "config": {"l1": 0.0, "l2": 1.000000013351432e-10}}, "bias_regularizer": null, "activity_regularizer": null, "kernel_constraint": null, "bias_constraint": null}, "name": "pattern_2", "inbound_nodes": [[["pattern_1", 0, 0, {}]]]}, {"class_name": "Conv1D", "config": {"name": "pattern_3", "trainable": true, "dtype": "float32", "filters": 16, "kernel_size": [4], "strides": [1], "padding": "same", "data_format": "channels_last", "dilation_rate": [1], "activation": "relu", "use_bias": true, "kernel_initializer": {"class_name": "VarianceScaling", "config": {"scale": 1.0, "mode": "fan_in", "distribution": "truncated_normal", "seed": null}}, "bias_initializer": {"class_name": "Zeros", "config": {}}, "kernel_regularizer": {"class_name": "L1L2", "config": {"l1": 0.0, "l2": 1.000000013351432e-10}}, "bias_regularizer": null, "activity_regularizer": null, "kernel_constraint": null, "bias_constraint": null}, "name": "pattern_3", "inbound_nodes": [[["pattern_2", 0, 0, {}]]]}, {"class_name": "Conv1D", "config": {"name": "pattern_4", "trainable": true, "dtype": "float32", "filters": 16, "kernel_size": [4], "strides": [1], "padding": "same", "data_format": "channels_last", "dilation_rate": [1], "activation": "relu", "use_bias": true, "kernel_initializer": {"class_name": "VarianceScaling", "config": {"scale": 1.0, "mode": "fan_in", "distribution": "truncated_normal", "seed": null}}, "bias_initializer": {"class_name": "Zeros", "config": {}}, "kernel_regularizer": {"class_name": "L1L2", "config": {"l1": 0.0, "l2": 1.000000013351432e-10}}, "bias_regularizer": null, "activity_regularizer": null, "kernel_constraint": null, "bias_constraint": null}, "name": "pattern_4", "inbound_nodes": [[["pattern_3", 0, 0, {}]]]}], "input_layers": [["hist", 0, 0]], "output_layers": [["pattern_4", 0, 0]]}, "keras_version": "2.2.4-tf", "backend": "tensorflow"}
