+++ PretrainedModels/patternModelReduced.json
@@ -1,0 +1,1 @@
+{"class_name": "Model", "config": {"name": "model_2", "layers": [{"class_name": "InputLayer", "config": {"batch_input_shape": [null, 256, 1], "dtype": "float32", "sparse": false, "ragged": false, "name": "hist"}, "name": "hist", "inbound_nodes": []}, {"class_name": "Conv1D", "config": {"name": "pattern_1", "trainable": true, "dtype": "float32", "filters": 1, "kernel_size": [3], "strides": [1], "padding": "same", "data_format": "channels_last", "dilation_rate": [1], "activation": "relu", "use_bias": false, "kernel_initializer": {"class_name": "GlorotUniform", "config": {"seed": null}}, "bias_initializer": {"class_name": "Zeros", "config": {}}, "kernel_regularizer": null, "bias_regularizer": null, "activity_regularizer": null, "kernel_constraint": null, "bias_constraint": null}, "name": "pattern_1", "inbound_nodes": [[["hist", 0, 0, {}]]]}], "input_layers": [["hist", 0, 0]], "output_layers": [["pattern_1", 0, 0]]}, "keras_version": "2.2.4-tf", "backend": "tensorflow"}
