+++ Train/vtx/nn/E2EDiffArgMax.py
@@ -3,17 +3,18 @@ import tensorflow_probability as tfp
 import vtx
 import numpy
 
-class E2Ecomparekernel():
+class E2EDiffArgMax():
     def __init__(self,
         nbins=256,
         ntracks=200, 
-        nweightfeatures=2,
-        nfeatures=10, 
+        nweightfeatures=1,
+        nfeatures=1, 
         nweights=1, 
-        npattern=16,
+        npattern=4,
         nlatent=0, 
         activation='relu',
-        regloss=1e-10
+        regloss=1e-10,
+        temperature=1e2
     ):
         self.nbins = nbins
         self.ntracks = ntracks
@@ -23,17 +24,20 @@ class E2Ecomparekernel():
         self.npattern = npattern
         self.nlatent = nlatent
         self.activation = activation
+
+        self.temperature = temperature
         
         self.inputWeightFeatures = tf.keras.layers.Input(shape=(self.ntracks,self.nweightfeatures),name='input_weight_features')
         self.inputTrackFeatures = tf.keras.layers.Input(shape=(self.ntracks,self.nfeatures),name='input_PV_track_features')
         self.inputTrackZ0 = tf.keras.layers.Input(shape=(self.ntracks),name='input_track_z0')
-        
+
         self.weightLayers = []
         for ilayer,nodes in enumerate([10,10]):
             self.weightLayers.extend([
                 tf.keras.layers.Dense(
                     nodes,
                     activation=self.activation,
+                    trainable=True,
                     kernel_initializer='lecun_normal',
                     kernel_regularizer=tf.keras.regularizers.l2(regloss),
                     name='weight_'+str(ilayer+1)
@@ -47,11 +51,11 @@ class E2Ecomparekernel():
                 self.nweights,
                 activation='relu', #need to use relu here to remove negative weights
                 kernel_initializer='lecun_normal',
+                trainable=True,
                 kernel_regularizer=tf.keras.regularizers.l2(regloss),
                 name='weight_final'
             ),
         )
-
         
         self.kdeLayer = vtx.nn.KDELayer(
             nbins=self.nbins,
@@ -62,10 +66,7 @@ class E2Ecomparekernel():
         
         self.patternConvLayers = []
         for ilayer,(filterSize,kernelSize) in enumerate([
-            [self.npattern,4],
-            [self.npattern,4],
-            [self.npattern,4],
-            [self.npattern,4],
+            [1,3],
         ]):
             self.patternConvLayers.append(
                 tf.keras.layers.Conv1D(
@@ -73,32 +74,29 @@ class E2Ecomparekernel():
                     kernelSize,
                     padding='same',
                     activation=self.activation,
-                    kernel_initializer='lecun_normal',
-                    kernel_regularizer=tf.keras.regularizers.l2(regloss),
+                    trainable=True,
+                    use_bias= False,
                     name='pattern_'+str(ilayer+1)
                 )
             )
-            
-            
-        self.positionConvLayers = []
-        for ilayer,(filterSize,kernelSize,strides) in enumerate([
-            [16,1,1],
-            [16,1,1],
-            [8,16,1],
-            [8,1,1],
-        ]):
-            self.positionConvLayers.append(
-                tf.keras.layers.Conv1D(
-                    filterSize,
-                    kernelSize,
-                    strides=strides,
-                    padding='same',
-                    activation=self.activation,
-                    kernel_initializer='lecun_normal',
-                    kernel_regularizer=tf.keras.regularizers.l2(regloss),
-                    name='position_'+str(ilayer+1)
+
+        
+
+        self.softMaxLayer = tf.keras.layers.Softmax(axis=1)
+
+        self.binWeightLayer = tf.keras.layers.Dense(
+                    self.nbins,
+                    activation='linear',
+                    trainable=False,
+                    use_bias= False,
+                    name='Bin_weight'
                 )
-            )
+
+        self.ArgMaxLayer = vtx.nn.BintoVertex(
+            nbins=self.nbins,
+            start=-15,
+            end=15
+        )
 
         self.pvDenseLayers = [
             tf.keras.layers.Dense(
@@ -132,15 +130,16 @@ class E2Ecomparekernel():
                 name='association_final'
             )
         ])
-        
+
         self.tiledTrackDimLayer = tf.keras.layers.Lambda(lambda x: tf.reshape(tf.tile(x,[1,self.ntracks]),[-1,self.ntracks,x.shape[1]]),name='tiled_track_dim')
-        
+
+                
     def applyLayerList(self, inputs, layerList):
         outputs = inputs
         for layer in layerList:
             outputs = layer(outputs)
         return outputs
-        
+
     def createWeightModel(self):
         weightInput = tf.keras.layers.Input(shape=(self.nweightfeatures),name="weight")
         weights = self.applyLayerList(weightInput,self.weightLayers)
@@ -148,15 +147,16 @@ class E2Ecomparekernel():
     
     def createPatternModel(self):
         histInput = tf.keras.layers.Input(shape=(self.nbins,self.nweights),name="hist")
-        pattern = self.applyLayerList(histInput,self.patternConvLayers)
-        return tf.keras.Model(inputs=[histInput],outputs=[pattern])
-    
+        convs = self.applyLayerList(histInput,self.patternConvLayers)
+        return tf.keras.Model(inputs=[histInput],outputs=[convs])
+
     def createPositionModel(self):
-        patternInput = tf.keras.layers.Input(shape=(self.npattern,self.nbins),name="pattern")
-        positionConv = self.applyLayerList(patternInput,self.positionConvLayers)
-        flattened = tf.keras.layers.Flatten()(positionConv)
-        pvFeatures = self.applyLayerList(flattened,self.pvDenseLayers)
-        return tf.keras.Model(inputs=[patternInput],outputs=[pvFeatures])
+        convsInput = tf.keras.layers.Input(shape=(self.nbins),name="conv")
+        temp = tf.keras.layers.Lambda(lambda x: x / self.temperature)
+        softmax = self.softMaxLayer(temp)
+        binweight = self.binWeightLayer(softmax)
+        argmax = self.ArgMaxLayer(binweight)
+        return tf.keras.Model(inputs=[convsInput],outputs=[argmax])
     
     def createAssociationModel(self):
         assocInput = tf.keras.layers.Input(shape=(self.nfeatures+1+self.nlatent),name="assoc")
@@ -164,16 +164,17 @@ class E2Ecomparekernel():
         return tf.keras.Model(inputs=[assocInput],outputs=[assocProbability])
         
     def createE2EModel(self):
+        
         weights = self.applyLayerList(self.inputWeightFeatures,self.weightLayers)
         hists = self.kdeLayer([self.inputTrackZ0,weights])
-        pattern = self.applyLayerList(hists,self.patternConvLayers)
-        
-        permuted = tf.keras.layers.Lambda(lambda x: tf.transpose(x,[0,2,1]))(pattern)
-        positionConv = self.applyLayerList(permuted,self.positionConvLayers)
-        flattened = tf.keras.layers.Flatten()(positionConv)
-        
-        pvFeatures = self.applyLayerList(flattened,self.pvDenseLayers)
-        
+        convs = self.applyLayerList(hists,self.patternConvLayers)
+        temp = tf.keras.layers.Lambda(lambda x: x / 1e-2)(convs)
+        softmax = self.softMaxLayer(temp)
+        binweight = self.binWeightLayer(softmax)
+        argmax = self.ArgMaxLayer(binweight)
+
+        pvFeatures = self.applyLayerList(argmax,self.pvDenseLayers)
+
         if self.nlatent>0:
             pvPosition,latentFeatures = tf.keras.layers.Lambda(lambda x: [x[:,0:1],x[:,1:]],name='split_latent')(pvFeatures)
         else:
@@ -181,20 +182,20 @@ class E2Ecomparekernel():
         
         z0Diff = tf.keras.layers.Lambda(lambda x: tf.stop_gradient(tf.expand_dims(tf.abs(x[0]-x[1]),2)),name='z0_diff')([self.inputTrackZ0,pvPosition])
         
-        assocFeatures = [self.inputTrackFeatures,z0Diff]
-        if self.nlatent>0:
-            assocFeatures.append(self.tiledTrackDimLayer(latentFeatures))
-            
-            
-        assocFeatures = tf.keras.layers.Concatenate(axis=2,name='association_features')(assocFeatures)
+        assocFeatures = [self.inputTrackFeatures,z0Diff]   
 
-        assocProbability = self.applyLayerList(assocFeatures,self.assocLayers)
+        if self.nlatent>0:
+            assocFeatures.append(self.tiledTrackDimLayer(latentFeatures))  
+            
+        assocFeat = tf.keras.layers.Concatenate(axis=2,name='association_features')(assocFeatures)
+
+        assocProbability = self.applyLayerList(assocFeat,self.assocLayers)
         
         model = tf.keras.Model(
             inputs=[self.inputTrackZ0,self.inputWeightFeatures,self.inputTrackFeatures],
             outputs=[pvPosition,assocProbability,weights]
         )
-        
+
         def q90loss(w):
             wq90 = tfp.stats.percentile(
                 w,
@@ -204,5 +205,5 @@ class E2Ecomparekernel():
             )
             return tf.reduce_mean(0.1*tf.square(wq90-1.))
         
-        model.add_loss(tf.keras.layers.Lambda(q90loss)(weights))
+        #model.add_loss(tf.keras.layers.Lambda(q90loss)(weights))
         return model
