+++ Train/vtx/nn/__init__.py
@@ -1,5 +1,8 @@
 from .KDELayer import *
+from .BintoVertex import *
 from .E2ERef import *
 from .E2Ecomparekernel import *
 from .E2Equartiles import *
 from .E2EReduced import *
+from .E2EFH import *
+from .E2EFHtrainable import *
